# ARROW-15440: [Go] Implement 'unpack_bool' with Arm64 GoLang Assembly