Benchmark:
Before:
```
BenchmarkPlainEncodingBoolean/len_1024-64 292210  4084 ns/op 250.74 MB/s
BenchmarkPlainEncodingBoolean/len_2048-64 179871  6669 ns/op 307.10 MB/s
BenchmarkPlainEncodingBoolean/len_4096-64 101024 11895 ns/op 344.35 MB/s
BenchmarkPlainEncodingBoolean/len_8192-64  53070 22599 ns/op 362.49 MB/s
BenchmarkPlainEncodingBoolean/len_16384-64 26664 45060 ns/op 363.60 MB/s
BenchmarkPlainEncodingBoolean/len_32768-64 13353 89844 ns/op 364.72 MB/s
BenchmarkPlainEncodingBoolean/len_65536-64  6663179439 ns/op 365.23 MB/s
```

After:
```
BenchmarkPlainEncodingBoolean/len_1024-64 414085  2898 ns/op 353.32 MB/s
BenchmarkPlainEncodingBoolean/len_2048-64 217814  5485 ns/op 373.39 MB/s
BenchmarkPlainEncodingBoolean/len_4096-64 112413 10678 ns/op 383.60 MB/s
BenchmarkPlainEncodingBoolean/len_8192-64  57085 21012 ns/op 389.87 MB/s
BenchmarkPlainEncodingBoolean/len_16384-64 28668 41854 ns/op 391.46 MB/s
BenchmarkPlainEncodingBoolean/len_32768-64 14356 83565 ns/op 392.13 MB/s
BenchmarkPlainEncodingBoolean/len_65536-64  7149166811 ns/op 392.88 MB/s
```

Get 7% ~ 40% performance uplift.

Tests in parquet/internal/encoding: 
```
ARM_ENABLE_EXT=NEON go test                                                                                  
PASS
ok      github.com/apache/arrow/go/v7/parquet/internal/encoding 9.949s
```